tests/wpt/web-platform-tests/css/work-in-progress/opera/animations/animation_shorthand_inline.html

<!DOCTYPE html>
<html>
<head>
   <title>CSS Animations - animation shorthand and inline element</title>
   <link rel="author" title="Opera Software ASA" href="http://www.opera.com/">
   <link rel="help" href="https://drafts.csswg.org/css3-animations/#the-animation-shorthand-property-">
   <meta name="flags" content="dom">
   <meta name="assert" content="The animation shorthand applies to inline
   elements">
   <script src="/resources/testharness.js"></script>
   <script src="/resources/testharnessreport.js"></script>
   <style type="text/css">
      @keyframes disappear {
         from {
            opacity: 1;
         }
         to {
            opacity: 0;
         }
      }
      span {
         color: blue;
      }
      .test {
         animation: disappear 1s 400ms steps(2) forwards;
      }
   </style>
   <script type="text/javascript">
      setup({explicit_done:true});
      onload = function() {
         var testName = document.querySelector('title').firstChild.data;
         var span = document.querySelector('span');
         var log = document.querySelectorAll('p')[1].firstChild;
         var style = getComputedStyle(span, null);
         var passed = false;
         setTimeout(function() {
            passed = (style.getPropertyValue('opacity') == '1');
         }, 200);
         setTimeout(function() {
            passed = (passed && style.getPropertyValue('opacity') == '0.5');
         }, 1100);
         setTimeout(function() {
            passed = (passed && style.getPropertyValue('opacity') == '0');
            test(function(){assert_true(passed);}, testName); done();
            log.data = (passed ? 'PASS if the animation occurred.' : 'FAIL');
         }, 1600);
         span.className = 'test';
      }
   </script>
   <style>iframe{display:none}</style>
</head>
<body>
   <p>The blue word below should become partially transparent, then completely
   transparent (i.e. invisible).</p>
   <p>Testing...</p>
   <div>normal <span>DISAPPEAR</span> normal</div>
</body>
</html>
